sh: forward runlet signals to process group
Forward runlet signals generated using the "signal" command as Unix
signals to the container process group.

Note the handling of signals is not consistent between commands and
pipes. pipes send themselves a (fake) "SIGALRM" to force a scan of the
process mailbox, since the process is blocked processing the stream.

    sh "cat" # standalone command
    query "true" | sh "cat" # pipe

    signal 1234 "term"

For the standalone command, a SIGTERM would be sent to cat and to the
container init.

For the pipe, if the signal is not handled by a previous command in the
chain, the signal will force the process mailbox to be rescanned but
otherwise ignored.