sh: pass through SIGALRM for pipe
Sh.exec/2 supports piping data from a stream resource. Pipes are used to
stream events into a containerized process. For example, selecting
failed events for a host and running network diagnostics.

The Sh.exec/2 implementation blocks on the stream. To force polling the
process mailbox, Sh.exec/2 sends itself a SIGALRM which the stream
resource passes through as an element in the stream.

Sh.exec/1 passes through signals and forwards the signals to the
container process group. Piping the output of a container to a container
would terminate the stream resource.

-   Sh.exec/1: pass through SIGALRM

-   Sh.exec/2: add tests