pipe: handle full pipe buffer
In exec mode, prx returns write failures to stdin asynchronously as
messages.

The pipe fd is non-blocking. Writes from prx go to the pipe buffer. Even
if the subprocess is not reading from the pipe, buffered writes will
return ok (0).  If the pipe buffer fills up, the write will return EAGAIN.

If the subprocess is not reading stdin or is unable to handle the input
rate, signal the process and exit.