Flush output from container
If a containerized process is running when the stream halts, kill the
process and wait indefinitely for a termination signal or exit status.
On a loaded system with many processes on the run queue, it may take
some time for the process to shut down.

Once received, process any remaining messages in the mailbox before
exiting.

* reduce duplication by handling stream and pipes using a common
  function

* ensure all output from the Linux process is sent into the stream by
  shutting down the Linux process on stream halt and flushing the process
  mailbox

* if the Linux process has already exited, exit immediately after all
  messages are received

* if the Linux process was still running, terminate using SIGKILL and
  wait for the process to exit before halting the stream